Add hall off fame update back-end
Retrieved form data from html form to the view function. Also, put the previous values of the row to update to the input areas so that the users can change only the feature that they want. The view function only prints the retreived data along with previous yearid value. Database connections will be added